+++ packages/client/src/Call.ts
@@ -285,7 +285,7 @@ export class Call {
           currentUserId &&
           metadata.blocked_user_ids.includes(currentUserId)
         ) {
-          this.logger('info', 'Leaving call bacause of being blocked');
+          this.logger('info', 'Leaving call because of being blocked');
           await this.leave();
         }
       }),
@@ -1602,7 +1602,7 @@ export class Call {
   /**
    * Sends a custom event to all call participants.
    *
-   * @param event the event to send.
+   * @param payload the payload to send.
    */
   sendCustomEvent = async (payload: { [key: string]: any }) => {
     return this.streamClient.post<SendEventResponse, SendEventRequest>(