+++ packages/react-native-sdk/src/providers/StreamCall.tsx
@@ -1,17 +1,12 @@
 import { StreamCallProvider } from '@stream-io/video-react-bindings';
 import React, { PropsWithChildren } from 'react';
 import { Call } from '@stream-io/video-client';
-import {
-  useAndroidKeepCallAliveEffect,
-  useCallCycleEffect,
-  usePermissionRequest,
-} from '../hooks';
+import { useAndroidKeepCallAliveEffect, usePermissionRequest } from '../hooks';
 import { useIosCallkeepEndEffect } from '../hooks/useIosCallkeepEndEffect';
 import { MediaStreamManagement } from './MediaStreamManagement';
 
 export type StreamCallProps = {
   call: Call;
-  callCycleHandlers?: CallCycleHandlersType;
 };
 /**
  * StreamCall is a wrapper component that orchestrates the call life cycle logic and
@@ -22,7 +17,6 @@ export type StreamCallProps = {
  */
 export const StreamCall = ({
   call,
-  callCycleHandlers = {},
   children,
 }: PropsWithChildren<StreamCallProps>) => {
   return (
@@ -31,61 +25,12 @@ export const StreamCall = ({
         <PermissionRequest />
         <AndroidKeepCallAlive />
         <IosInformCallkeepCallEnd />
-        <CallCycleLogicsWrapper callCycleHandlers={callCycleHandlers}>
-          {children}
-        </CallCycleLogicsWrapper>
+        {children}
       </MediaStreamManagement>
     </StreamCallProvider>
   );
 };
 
-/**
- * Exclude types from documentation site, but we should still add doc comments
- * @internal
- */
-export type CallCycleHandlersType = {
-  /**
-   * Handler called after a call is joined. Mostly used for navigation and related actions.
-   */
-  onCallJoined?: () => void;
-  /**
-   * Handler called after a callee receives a call. Mostly used for navigation and related actions.
-   */
-  onCallIncoming?: () => void;
-  /**
-   * Handler called after a call is hung up by the caller. Mostly used for navigation and cleanup actions.
-   */
-  onCallHungUp?: () => void;
-  /**
-   * Handler called after a caller initiates a call. Mostly used for navigation and related actions.
-   */
-  onCallOutgoing?: () => void;
-  /**
-   * Handler called after a call is rejected. Mostly used for navigation and cleanup actions.
-   */
-  onCallRejected?: () => void;
-  /**
-   * Handler called when the call is in joining state. Mostly used for navigation and related actions.
-   */
-  onCallJoining?: () => void;
-};
-
-/**
- * Exclude types from documentaiton site, but we should still add doc comments
- * @internal
- */
-export type CallCycleLogicsWrapperProps = {
-  callCycleHandlers: CallCycleHandlersType;
-};
-
-export const CallCycleLogicsWrapper = ({
-  callCycleHandlers,
-  children,
-}: PropsWithChildren<CallCycleLogicsWrapperProps>) => {
-  useCallCycleEffect(callCycleHandlers);
-
-  return <>{children}</>;
-};
 /**
  * This is a renderless component that is used to handler the permission requests using the usePermissionRequest hook.
  * usePermissionRequest needs to be called as a child of StreamCallProvider.